refactor(eval-runtime)!: 命令式 builder 归入 canonical façade，advanced 纯化为宿主装配 SPI
BREAKING-API：eval-runtime/advanced 的公开表面收缩并改名。

- 命令式 builder（createExactMatchDefinition、createPairedComparisonDefinition、
  createMeasurementPolicy、createExactMatchEvaluator 及 identity、createRubricJudge*
  全套、rubricJudgeInstrumentId 与相关类型）从 eval-runtime/advanced 迁到
  eval-runtime 主路径（包根 façade）。advanced 只保留宿主装配 SPI（Runtime 装配、
  adapter、conformance、execution-control SPI、identity、clock）。
- advanced 的 EvaluationEventConsumptionError 改名为 HostEvaluationEventConsumptionError
  （带 cause 的 runEvaluation 宿主入口错误），与 façade 脱敏版 EvaluationEventConsumptionError
  区分，消除同名不同类的 instanceof 陷阱。
- 补出 advanced 缺失的 MockInterceptionInput/MockInterceptionPlan，使四种
  execution-control 的 SPI 导出一致。
- façade 词汇 guard 收窄 target 禁令：只禁裸 target/targets（artifact 同义词），
  豁免 XxxTarget/targetId 等正统 Core Target 概念。

迁移：builder 与相关类型改从 oh-my-knowledge/eval-runtime（或包根）导入；
advanced 侧捕获事件消费错误的代码改用 HostEvaluationEventConsumptionError。
