+++ docs/guides/platform-host-integration.md
@@ -159,7 +159,7 @@ The export list, per-code meanings and drift rules are in the [Reference Executo
 
 Audit-grade write-back combines two fields:
 
-- Declare the delivery mode in `EvaluateInput.policy.eventDelivery` when preparing or evaluating (the shape is `MeasurementEventDeliveryInput`, nameable from `oh-my-knowledge/eval-runtime/advanced`; on the façade you only write the literal): `writerMode` is `disabled` (the default), `optional`, or `required`, and `writerFailureMode` is `ignore` or `fail-run`. The pairing is strictly validated: `disabled` accepts only `ignore`, `required` accepts only `fail-run` and defaults to it, and `optional` defaults to `ignore`.
+- Declare the delivery mode in `EvaluateInput.policy.eventDelivery` when preparing or evaluating (the shape is `MeasurementEventDeliveryInput`, nameable from `oh-my-knowledge/eval-runtime`; on the façade you only write the literal): `writerMode` is `disabled` (the default), `optional`, or `required`, and `writerFailureMode` is `ignore` or `fail-run`. The pairing is strictly validated: `disabled` accepts only `ignore`, `required` accepts only `fail-run` and defaults to it, and `optional` defaults to `ignore`.
 - Pass a writer in `EvaluationRunOptions.eventWriter` to persist events one by one in order; the event shape is the published schema `evaluation-event.schema.json` (resolvable with `resolveEvaluationCoreJsonSchema`). **Completeness is guaranteed only by `required`**: it turns a failed write into a failed run, so a hole in the process record is impossible to mistake for a complete one. `optional` + `ignore` is best-effort — the first failed write silently stops durable delivery for that stage, the run still completes, and nothing in the result reports the truncation, so never use it where the record is the deliverable.
 
 ```ts
