+++ test/eval-core/embedded-package.test.ts
@@ -235,11 +235,11 @@ const assert = require('node:assert/strict');
   assert.equal(evalRuntime.RUNTIME_CHECK_RESULT_SCHEMA_VERSION, 'omk.runtime-check-result/v1');
   assert.equal(evalRuntime.createExecutorFnAdapter, undefined);
   assert.equal(evalRuntime.createJsonExecutorAdapter, undefined);
-  assert.equal(evalRuntime.createRubricJudgeKit, undefined);
   assert.equal(evalRuntime.runEvaluation, undefined);
+  assert.equal(typeof evalRuntime.createRubricJudgeKit, 'function');
   assert.equal(typeof evalRuntimeAdvanced.createEvaluationRuntime, 'function');
   assert.equal(typeof evalRuntimeAdvanced.createJsonExecutorAdapter, 'function');
-  assert.equal(typeof evalRuntimeAdvanced.createRubricJudgeKit, 'function');
+  assert.equal(evalRuntimeAdvanced.createRubricJudgeKit, undefined);
   assert.equal(typeof evalRuntimeAdvanced.runEvaluation, 'function');
   assert.equal(typeof evalRuntimeAdvanced.createExecutorFnAdapter, 'function');
   assert.equal(typeof evalRuntimeAdvanced.createSameProcessExecutorAdapter, 'function');
