+++ test/eval-runtime/foundation.test.ts
@@ -4,18 +4,20 @@ import { createEvaluationEngine } from '../../src/eval-core/index.js';
 import {
   assertExecutorConformance,
   createEvaluationRuntime,
-  createExactMatchDefinition,
-  createExactMatchEvaluator,
   createInvokeExecutorIdentity,
   createJsonExecutorAdapter,
-  createMeasurementPolicy,
-  createPairedComparisonDefinition,
   runExecutorConformance,
 } from '../../src/eval-runtime/advanced.js';
 import {
   createExecutorFnAdapter,
   type ExecResult,
 } from '../../src/eval-runtime/advanced.js';
+import {
+  createExactMatchDefinition,
+  createExactMatchEvaluator,
+  createMeasurementPolicy,
+  createPairedComparisonDefinition,
+} from '../../src/eval-runtime/index.js';
 
 function result(output: string): ExecResult {
   return {
