+++ test/eval-runtime/public-api-contract.ts
@@ -1,0 +1,432 @@
+import { existsSync, readFileSync } from 'node:fs';
+import { resolve } from 'node:path';
+import ts from 'typescript';
+
+/**
+ * eval-runtime 公开 API allowlist：发布契约的单一来源。
+ *
+ * 被两个测试文件共享，避免双轨漂移：
+ * - public-api-source.test.ts 用 src 提取导出名集合做快速反馈（不依赖 build）。
+ * - public-api.test.ts 用 dist 提取运行时导出与 .d.ts 值/类型分类做发布契约锁定。
+ */
+export const PUBLIC_API = {
+  'eval-runtime': {
+    entry: 'index',
+    values: [
+      'EVALUATION_RESULT_MEDIA_TYPE',
+      'EXACT_MATCH_EVALUATOR_IMPLEMENTATION_ID',
+      'EXECUTED_EVALUATION_MEDIA_TYPE',
+      'EvaluationConfigurationError',
+      'EvaluationEventConsumptionError',
+      'EvaluationResultStoreError',
+      'ExecutedEvaluationStoreError',
+      'MOCK_INTERCEPTION_PLAN_MEDIA_TYPE',
+      'RUNTIME_CHECK_RESULT_SCHEMA_VERSION',
+      'assessComparability',
+      'checkContentStore',
+      'checkExecutor',
+      'checkRuntime',
+      'createExactMatchDefinition',
+      'createExactMatchEvaluator',
+      'createExactMatchEvaluatorIdentity',
+      'createMeasurementPolicy',
+      'createPairedComparisonDefinition',
+      'createRubricJudgeCriterion',
+      'createRubricJudgeEvaluationContext',
+      'createRubricJudgeEvaluator',
+      'createRubricJudgeEvaluatorDefinition',
+      'createRubricJudgeEvaluatorIdentity',
+      'createRubricJudgeEvaluatorRegistration',
+      'createRubricJudgeInstrument',
+      'createRubricJudgeKit',
+      'createRubricJudgeMetricDefinition',
+      'createRubricJudgeRegistration',
+      'createRubricJudgeRuntimeConfig',
+      'evaluate',
+      'evaluateSeries',
+      'executeEvaluation',
+      'loadEvaluationResult',
+      'loadExecutedEvaluation',
+      'prepareEvaluation',
+      'prepareEvaluationSeries',
+      'reanalyze',
+      'redecide',
+      'rescore',
+      'rubricJudgeInstrumentId',
+      'saveEvaluationResult',
+      'saveExecutedEvaluation',
+      'scoreExecutedEvaluation',
+    ],
+    types: [
+      'AbstentionEvaluator',
+      'AbstentionMetricIds',
+      'AllowedToolsInput',
+      'AllowedToolsPlan',
+      'AnalysisRequest',
+      'Artifact',
+      'ArtifactKind',
+      'ArtifactSource',
+      'AssessComparabilityInput',
+      'AttemptBudgetScope',
+      'BudgetPolicy',
+      'BudgetScope',
+      'CacheConformanceCheck',
+      'CacheConformanceProbeInput',
+      'CacheConformanceResult',
+      'CachePolicy',
+      'CacheRuntimeCheckInput',
+      'CacheRuntimeCheckResult',
+      'Clock',
+      'CohortFilter',
+      'Comparison',
+      'ComparisonFamilyMember',
+      'CompositeAggregation',
+      'CompositeMetricComponent',
+      'ContentDescriptor',
+      'ContentResolver',
+      'ContentStore',
+      'ContentStoreCheckInput',
+      'ContentStoreCheckResult',
+      'ContentStoreConformanceCheck',
+      'ContentStoreRequest',
+      'ContentStoreRuntimeCheckInput',
+      'ContentStoreRuntimeCheckResult',
+      'ContentValue',
+      'CreateExactMatchEvaluatorInput',
+      'CreateRubricJudgeEvaluatorInput',
+      'CreateRubricJudgeKitInput',
+      'CustomEvaluator',
+      'CustomEvaluatorBinding',
+      'CustomEvaluatorContent',
+      'CustomEvaluatorInvocation',
+      'CustomEvaluatorResult',
+      'Dataset',
+      'Decision',
+      'EvaluateInput',
+      'EvaluationCache',
+      'EvaluationCacheEntry',
+      'EvaluationComparabilityAssessment',
+      'EvaluationComparabilitySubject',
+      'EvaluationEventWriter',
+      'EvaluationExecutor',
+      'EvaluationInfrastructure',
+      'EvaluationResult',
+      'EvaluationResultVerification',
+      'EvaluationResultVerificationRequest',
+      'EvaluationResultVerifier',
+      'EvaluationRunOptions',
+      'EvaluationRuntimeTarget',
+      'EvaluationSeriesInput',
+      'EvaluationSeriesMemberResult',
+      'EvaluationSeriesResult',
+      'EvaluationSeriesRunOptions',
+      'EvaluationSeriesStability',
+      'EvaluationSeriesStabilityResult',
+      'EvaluationSeriesWorkEstimate',
+      'EvaluationWorkEstimate',
+      'Evaluator',
+      'EvaluatorConformanceCheck',
+      'EvaluatorConformanceProbeInput',
+      'EvaluatorConformanceProbeSources',
+      'EvaluatorConformanceResult',
+      'EvaluatorRuntimeCheckInput',
+      'EvaluatorRuntimeCheckResult',
+      'EventObserver',
+      'EvidencePolicy',
+      'ExactMatchDefinitionBuilderInput',
+      'ExactMatchEvaluator',
+      'ExactMatchTarget',
+      'ExecutedEvaluation',
+      'ExecutedEvaluationVerification',
+      'ExecutedEvaluationVerificationRequest',
+      'ExecutedEvaluationVerifier',
+      'ExecutionCache',
+      'ExecutionCacheEntry',
+      'Executor',
+      'ExecutorCapabilities',
+      'ExecutorCheckInput',
+      'ExecutorCheckResult',
+      'ExecutorIdentityVerification',
+      'ExecutorIdentityVerificationRequest',
+      'ExecutorIdentityVerifier',
+      'ExecutorInvocation',
+      'ExecutorResult',
+      'ExecutorRuntimeCheckInput',
+      'ExecutorRuntimeCheckResult',
+      'ExecutorSession',
+      'ExecutorSessionAttempt',
+      'ExecutorSessionContext',
+      'Experiment',
+      'FailurePolicy',
+      'FamilyDecisionCriterion',
+      'InvokeExecutor',
+      'Judge',
+      'JudgeConformanceCheck',
+      'JudgeConformanceProbeCase',
+      'JudgeConformanceProbeInput',
+      'JudgeConformanceResult',
+      'JudgeRuntimeCheckInput',
+      'JudgeRuntimeCheckResult',
+      'LoadEvaluationResultInput',
+      'LoadExecutedEvaluationInput',
+      'McpConfigAccess',
+      'McpConfigDescriptor',
+      'McpConfigInput',
+      'McpConfigLease',
+      'McpConfigOpenRequest',
+      'McpConfigPlan',
+      'McpConfigProvider',
+      'MeasurementAttemptBudgetScopeInput',
+      'MeasurementBudgetPolicyInput',
+      'MeasurementBudgetScopeInput',
+      'MeasurementCachePolicyInput',
+      'MeasurementEventDeliveryInput',
+      'MeasurementEvidencePolicyInput',
+      'MeasurementFailurePolicyInput',
+      'MeasurementPolicyBuilderInput',
+      'MeasurementProviderCostLimitInput',
+      'MeasurementRetryBackoffInput',
+      'MeasurementRetryPolicyInput',
+      'MeasurementRunBudgetScopeInput',
+      'MeasurementStagePolicyInput',
+      'Metric',
+      'MockInterceptionAccess',
+      'MockInterceptionDecision',
+      'MockInterceptionDescriptor',
+      'MockInterceptionInput',
+      'MockInterceptionLease',
+      'MockInterceptionOpenRequest',
+      'MockInterceptionPlan',
+      'MockInterceptionProvider',
+      'MockInterceptionRequest',
+      'PairedComparisonDefinitionBuilderInput',
+      'Policy',
+      'PreparedEvaluation',
+      'PreparedEvaluationPlan',
+      'PreparedEvaluationSeries',
+      'ProviderCostLimit',
+      'RetrievalEvaluator',
+      'RetrievalMetricIds',
+      'RetryBackoff',
+      'RetryPolicy',
+      'Rubric',
+      'RubricJudgeAggregation',
+      'RubricJudgeEvaluator',
+      'RubricJudgeEvaluatorBinding',
+      'RubricJudgeEvaluatorDefinitionBuilderInput',
+      'RubricJudgeKit',
+      'RubricJudgeMember',
+      'RunBudgetScope',
+      'RunStabilityValue',
+      'RuntimeCapabilityResolution',
+      'RuntimeCheckInput',
+      'RuntimeCheckKind',
+      'RuntimeCheckResult',
+      'RuntimeConformanceCheck',
+      'RuntimeContext',
+      'Sample',
+      'SamplingDesign',
+      'SaveEvaluationResultInput',
+      'SaveExecutedEvaluationInput',
+      'SessionExecutor',
+      'StagePolicy',
+      'ToolTrajectoryEvaluator',
+      'ToolTrajectoryMatchMode',
+      'Variant',
+      'VariantExecution',
+      'WorkspaceAccess',
+      'WorkspaceDescriptor',
+      'WorkspaceInput',
+      'WorkspaceLease',
+      'WorkspaceOpenRequest',
+      'WorkspacePlan',
+      'WorkspaceProvider',
+      'WorkspaceProviderConformanceCheck',
+      'WorkspaceProviderConformanceProbeInput',
+      'WorkspaceProviderConformanceResult',
+      'WorkspaceProviderRuntimeCheckInput',
+      'WorkspaceProviderRuntimeCheckResult',
+    ],
+  },
+  'eval-runtime/advanced': {
+    entry: 'advanced',
+    values: [
+      'DEFAULT_SUBPROCESS_COMMAND_MAX_OUTPUT_BYTES',
+      'EvaluationRuntimeAssemblyError',
+      'HostEvaluationEventConsumptionError',
+      'INVOKE_JSON_INPUT_SCHEMA',
+      'INVOKE_JSON_OUTPUT_SCHEMA',
+      'INVOKE_JSON_TRACE_SCHEMA',
+      'MOCK_INTERCEPTION_PLAN_MEDIA_TYPE',
+      'RuntimeConformanceError',
+      'SESSION_JSON_INPUT_SCHEMA',
+      'SESSION_JSON_OUTPUT_SCHEMA',
+      'SESSION_JSON_TRACE_SCHEMA',
+      'SUBPROCESS_COMMAND_EXCHANGE_SCHEMA_VERSION',
+      'assertExecutorConformance',
+      'createEvaluationRuntime',
+      'createExecutorFnAdapter',
+      'createInvokeExecutorIdentity',
+      'createJsonExecutorAdapter',
+      'createJsonSessionExecutorAdapter',
+      'createNodeEvaluationClock',
+      'createRuntimeIdentity',
+      'createSameProcessEvaluatorAdapter',
+      'createSameProcessExecutorAdapter',
+      'createSessionExecutorIdentity',
+      'createSubprocessCommandExecutor',
+      'runEvaluation',
+      'runExecutorConformance',
+    ],
+    types: [
+      'AllowedToolsInput',
+      'AllowedToolsPlan',
+      'CreateEvaluationRuntimeInput',
+      'CreateExecutorFnAdapterInput',
+      'CreateJsonExecutorAdapterInput',
+      'CreateJsonSessionExecutorAdapterInput',
+      'CreateSameProcessEvaluatorAdapterInput',
+      'CreateSameProcessExecutorAdapterInput',
+      'CreateSubprocessCommandExecutorInput',
+      'EvaluationEventObserver',
+      'EvaluationRuntimeSupportPorts',
+      'ExecResult',
+      'ExecutorConformanceProbeInput',
+      'ExecutorConformanceResult',
+      'ExecutorFn',
+      'ExecutorFnInputMapper',
+      'ExecutorFnResultMapper',
+      'ExecutorInput',
+      'InvokeExecutorIdentityDeclaration',
+      'JsonExecutorInvocation',
+      'JsonExecutorInvocationResult',
+      'JsonExecutorSession',
+      'JsonSessionExecutorAttempt',
+      'JsonSessionExecutorContext',
+      'McpConfigAccess',
+      'McpConfigDescriptor',
+      'McpConfigInput',
+      'McpConfigLease',
+      'McpConfigOpenRequest',
+      'McpConfigPlan',
+      'McpConfigProvider',
+      'MockInterceptionAccess',
+      'MockInterceptionDecision',
+      'MockInterceptionDescriptor',
+      'MockInterceptionInput',
+      'MockInterceptionLease',
+      'MockInterceptionOpenRequest',
+      'MockInterceptionPlan',
+      'MockInterceptionProvider',
+      'MockInterceptionRequest',
+      'OmkLlmJudgeEffort',
+      'OmkLlmJudgeInvocationPort',
+      'OmkLlmJudgeInvocationRequest',
+      'OmkLlmJudgeInvocationResult',
+      'RunEvaluationInput',
+      'RuntimeConformanceCheck',
+      'RuntimeIdentityDeclaration',
+      'RuntimePortRegistration',
+      'RuntimeValueParser',
+      'SameProcessEvaluatorImplementation',
+      'SameProcessExecutorImplementation',
+      'SameProcessOperationScope',
+      'SameProcessResourceLeaseAccess',
+      'SameProcessRunScope',
+      'SessionExecutorIdentityDeclaration',
+      'SubprocessCommandConfiguration',
+      'SubprocessCommandValueParser',
+      'WorkspaceAccess',
+      'WorkspaceDescriptor',
+      'WorkspaceInput',
+      'WorkspaceLease',
+      'WorkspaceOpenRequest',
+      'WorkspacePlan',
+      'WorkspaceProvider',
+    ],
+  },
+  'eval-runtime/contracts': {
+    entry: 'contracts',
+    values: [
+      'RUBRIC_JUDGE_BINDINGS',
+      'RUBRIC_JUDGE_CONTEXT_SCHEMA',
+      'RUBRIC_JUDGE_CONTEXT_SCHEMA_VERSION',
+      'RUBRIC_JUDGE_EVALUATOR_IMPLEMENTATION_ID',
+      'RUBRIC_JUDGE_EVIDENCE_SCHEMA',
+      'RUBRIC_JUDGE_EVIDENCE_SCHEMA_VERSION',
+      'RUBRIC_JUDGE_INSTRUMENT_SCHEMA',
+      'RUBRIC_JUDGE_INSTRUMENT_SCHEMA_VERSION',
+      'SOURCE_NEUTRAL_TRACE_SCHEMA_DESCRIPTOR',
+      'SOURCE_NEUTRAL_TRACE_SCHEMA_VERSION',
+      'SOURCE_NEUTRAL_TRACE_WITHOUT_MOCKS_SCHEMA_DESCRIPTOR',
+      'SourceNeutralMockStatsSchema',
+      'SourceNeutralTraceSchema',
+      'SourceNeutralTraceWithoutMocksSchema',
+      'attachSourceNeutralMockStats',
+      'parseSourceNeutralTrace',
+    ],
+    types: [
+      'RubricJudgeConfig',
+      'RubricJudgeCriterion',
+      'RubricJudgeInstrument',
+      'RubricJudgeRuntimeConfig',
+      'RubricJudgeTracePolicy',
+      'SourceNeutralMockStats',
+      'SourceNeutralTrace',
+    ],
+  },
+} as const;
+
+export type PublicApiSubpath = keyof typeof PUBLIC_API;
+
+/** 从 dist 的 .d.ts 提取命名导出，区分值导出与类型导出。 */
+export function declarationExports(entry: string): { values: string[]; types: string[] } {
+  const file = resolve(`dist/eval-runtime/${entry}.d.ts`);
+  if (!existsSync(file)) throw new Error(`缺少 ${file}；请先运行 yarn build。`);
+  const source = ts.createSourceFile(
+    file,
+    readFileSync(file, 'utf8'),
+    ts.ScriptTarget.Latest,
+    true,
+    ts.ScriptKind.TS,
+  );
+  const values: string[] = [];
+  const types: string[] = [];
+  for (const statement of source.statements) {
+    if (!ts.isExportDeclaration(statement)
+        || statement.exportClause === undefined
+        || !ts.isNamedExports(statement.exportClause)) continue;
+    for (const element of statement.exportClause.elements) {
+      (statement.isTypeOnly || element.isTypeOnly ? types : values).push(element.name.text);
+    }
+  }
+  return { values: values.sort(), types: types.sort() };
+}
+
+/**
+ * 从 src 的 .ts 入口提取命名导出的名字集合（不区分值/类型）。
+ *
+ * 与 declarationExports 共用 AST 解析，但读 src 而非 dist，因此无需 build。
+ * 值/类型分类仍由 dist 的 .d.ts 层校验；src 层只保证导出名齐全。
+ */
+export function sourceExportNames(entry: string): string[] {
+  const file = resolve(`src/eval-runtime/${entry}.ts`);
+  if (!existsSync(file)) throw new Error(`缺少 ${file}。`);
+  const source = ts.createSourceFile(
+    file,
+    readFileSync(file, 'utf8'),
+    ts.ScriptTarget.Latest,
+    true,
+    ts.ScriptKind.TS,
+  );
+  const names: string[] = [];
+  for (const statement of source.statements) {
+    if (!ts.isExportDeclaration(statement)
+        || statement.exportClause === undefined
+        || !ts.isNamedExports(statement.exportClause)) continue;
+    for (const element of statement.exportClause.elements) {
+      names.push(element.name.text);
+    }
+  }
+  return names.sort();
+}
