+++ test/eval-runtime/public-api.test.ts
@@ -1,8 +1,8 @@
-import { existsSync, readFileSync } from 'node:fs';
+import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
-import ts from 'typescript';
 import { z } from 'zod';
 import { describe, expect, it } from 'vitest';
+import { PUBLIC_API, declarationExports } from './public-api-contract.js';
 import type {
   AbstentionEvaluator,
   AbstentionMetricIds,
@@ -218,395 +218,6 @@ const publicCustomEvaluator = {
     },
   },
 } satisfies CustomEvaluator<{ actual: string }, { trim: boolean }>;
-
-const PUBLIC_API = {
-  'eval-runtime': {
-    entry: 'index',
-    values: [
-      'EvaluationConfigurationError',
-      'EvaluationEventConsumptionError',
-      'EVALUATION_RESULT_MEDIA_TYPE',
-      'EvaluationResultStoreError',
-      'EXECUTED_EVALUATION_MEDIA_TYPE',
-      'ExecutedEvaluationStoreError',
-      'MOCK_INTERCEPTION_PLAN_MEDIA_TYPE',
-      'RUNTIME_CHECK_RESULT_SCHEMA_VERSION',
-      'assessComparability',
-      'checkContentStore',
-      'checkExecutor',
-      'checkRuntime',
-      'evaluate',
-      'evaluateSeries',
-      'executeEvaluation',
-      'loadEvaluationResult',
-      'loadExecutedEvaluation',
-      'prepareEvaluation',
-      'prepareEvaluationSeries',
-      'reanalyze',
-      'redecide',
-      'rescore',
-      'saveEvaluationResult',
-      'saveExecutedEvaluation',
-      'scoreExecutedEvaluation',
-    ],
-    types: [
-      'AbstentionEvaluator',
-      'AbstentionMetricIds',
-      'AllowedToolsInput',
-      'AllowedToolsPlan',
-      'Artifact',
-      'ArtifactKind',
-      'ArtifactSource',
-      'AnalysisRequest',
-      'AssessComparabilityInput',
-      'AttemptBudgetScope',
-      'BudgetPolicy',
-      'BudgetScope',
-      'CachePolicy',
-      'CacheConformanceCheck',
-      'CacheConformanceProbeInput',
-      'CacheConformanceResult',
-      'CacheRuntimeCheckInput',
-      'CacheRuntimeCheckResult',
-      'Clock',
-      'CohortFilter',
-      'Comparison',
-      'ComparisonFamilyMember',
-      'CompositeAggregation',
-      'CompositeMetricComponent',
-      'ContentDescriptor',
-      'ContentResolver',
-      'ContentStore',
-      'ContentStoreCheckInput',
-      'ContentStoreCheckResult',
-      'ContentStoreConformanceCheck',
-      'ContentStoreRuntimeCheckInput',
-      'ContentStoreRuntimeCheckResult',
-      'ContentStoreRequest',
-      'ContentValue',
-      'CustomEvaluator',
-      'CustomEvaluatorBinding',
-      'CustomEvaluatorContent',
-      'CustomEvaluatorInvocation',
-      'CustomEvaluatorResult',
-      'Dataset',
-      'Decision',
-      'EvaluateInput',
-      'EvaluationExecutor',
-      'EvaluationComparabilityAssessment',
-      'EvaluationComparabilitySubject',
-      'EvaluationEventWriter',
-      'EvaluationInfrastructure',
-      'EvaluationCache',
-      'EvaluationCacheEntry',
-      'EvaluationRunOptions',
-      'EvaluationResult',
-      'EvaluationResultVerification',
-      'EvaluationResultVerificationRequest',
-      'EvaluationResultVerifier',
-      'EvaluationSeriesInput',
-      'EvaluationSeriesMemberResult',
-      'EvaluationSeriesResult',
-      'EvaluationSeriesRunOptions',
-      'EvaluationSeriesStability',
-      'EvaluationSeriesStabilityResult',
-      'EvaluationSeriesWorkEstimate',
-      'EvaluationWorkEstimate',
-      'ExecutedEvaluation',
-      'ExecutedEvaluationVerification',
-      'ExecutedEvaluationVerificationRequest',
-      'ExecutedEvaluationVerifier',
-      'EvidencePolicy',
-      'Evaluator',
-      'EvaluatorConformanceCheck',
-      'EvaluatorConformanceProbeInput',
-      'EvaluatorConformanceProbeSources',
-      'EvaluatorConformanceResult',
-      'EvaluatorRuntimeCheckInput',
-      'EvaluatorRuntimeCheckResult',
-      'EventObserver',
-      'ExactMatchEvaluator',
-      'FailurePolicy',
-      'FamilyDecisionCriterion',
-      'Executor',
-      'ExecutorCapabilities',
-      'ExecutorCheckInput',
-      'ExecutorCheckResult',
-      'ExecutorInvocation',
-      'ExecutorIdentityVerification',
-      'ExecutorIdentityVerificationRequest',
-      'ExecutorIdentityVerifier',
-      'ExecutorResult',
-      'ExecutorRuntimeCheckInput',
-      'ExecutorRuntimeCheckResult',
-      'ExecutionCache',
-      'ExecutionCacheEntry',
-      'ExecutorSession',
-      'ExecutorSessionAttempt',
-      'ExecutorSessionContext',
-      'Experiment',
-      'InvokeExecutor',
-      'Judge',
-      'JudgeConformanceCheck',
-      'JudgeConformanceProbeCase',
-      'JudgeConformanceProbeInput',
-      'JudgeConformanceResult',
-      'JudgeRuntimeCheckInput',
-      'JudgeRuntimeCheckResult',
-      'McpConfigAccess',
-      'McpConfigDescriptor',
-      'McpConfigInput',
-      'McpConfigLease',
-      'McpConfigOpenRequest',
-      'McpConfigPlan',
-      'McpConfigProvider',
-      'MockInterceptionAccess',
-      'MockInterceptionDecision',
-      'MockInterceptionDescriptor',
-      'MockInterceptionInput',
-      'MockInterceptionLease',
-      'MockInterceptionOpenRequest',
-      'MockInterceptionPlan',
-      'MockInterceptionProvider',
-      'MockInterceptionRequest',
-      'Metric',
-      'Policy',
-      'PreparedEvaluation',
-      'PreparedEvaluationPlan',
-      'PreparedEvaluationSeries',
-      'ProviderCostLimit',
-      'RetryBackoff',
-      'RetryPolicy',
-      'RetrievalEvaluator',
-      'RetrievalMetricIds',
-      'Rubric',
-      'RubricJudgeAggregation',
-      'RubricJudgeEvaluator',
-      'RubricJudgeMember',
-      'RuntimeConformanceCheck',
-      'RuntimeCheckInput',
-      'RuntimeCheckKind',
-      'RuntimeCheckResult',
-      'RuntimeContext',
-      'RuntimeCapabilityResolution',
-      'RunBudgetScope',
-      'RunStabilityValue',
-      'SaveEvaluationResultInput',
-      'SaveExecutedEvaluationInput',
-      'LoadExecutedEvaluationInput',
-      'Sample',
-      'SamplingDesign',
-      'SessionExecutor',
-      'LoadEvaluationResultInput',
-      'StagePolicy',
-      'ToolTrajectoryEvaluator',
-      'ToolTrajectoryMatchMode',
-      'Variant',
-      'VariantExecution',
-      'WorkspaceAccess',
-      'WorkspaceDescriptor',
-      'WorkspaceInput',
-      'WorkspaceLease',
-      'WorkspaceOpenRequest',
-      'WorkspacePlan',
-      'WorkspaceProvider',
-      'WorkspaceProviderConformanceCheck',
-      'WorkspaceProviderConformanceProbeInput',
-      'WorkspaceProviderConformanceResult',
-      'WorkspaceProviderRuntimeCheckInput',
-      'WorkspaceProviderRuntimeCheckResult',
-    ],
-  },
-  'eval-runtime/advanced': {
-    entry: 'advanced',
-    values: [
-      'EXACT_MATCH_EVALUATOR_IMPLEMENTATION_ID',
-      'DEFAULT_SUBPROCESS_COMMAND_MAX_OUTPUT_BYTES',
-      'EvaluationEventConsumptionError',
-      'EvaluationRuntimeAssemblyError',
-      'INVOKE_JSON_INPUT_SCHEMA',
-      'INVOKE_JSON_OUTPUT_SCHEMA',
-      'INVOKE_JSON_TRACE_SCHEMA',
-      'MOCK_INTERCEPTION_PLAN_MEDIA_TYPE',
-      'SESSION_JSON_INPUT_SCHEMA',
-      'SESSION_JSON_OUTPUT_SCHEMA',
-      'SESSION_JSON_TRACE_SCHEMA',
-      'SUBPROCESS_COMMAND_EXCHANGE_SCHEMA_VERSION',
-      'RuntimeConformanceError',
-      'assertExecutorConformance',
-      'createEvaluationRuntime',
-      'createExactMatchDefinition',
-      'createExactMatchEvaluator',
-      'createExactMatchEvaluatorIdentity',
-      'createExecutorFnAdapter',
-      'createInvokeExecutorIdentity',
-      'createJsonExecutorAdapter',
-      'createJsonSessionExecutorAdapter',
-      'createMeasurementPolicy',
-      'createNodeEvaluationClock',
-      'createPairedComparisonDefinition',
-      'createRubricJudgeCriterion',
-      'createRubricJudgeEvaluationContext',
-      'createRubricJudgeEvaluator',
-      'createRubricJudgeEvaluatorDefinition',
-      'createRubricJudgeEvaluatorIdentity',
-      'createRubricJudgeEvaluatorRegistration',
-      'createRubricJudgeInstrument',
-      'createRubricJudgeKit',
-      'createRubricJudgeMetricDefinition',
-      'createRubricJudgeRegistration',
-      'createRubricJudgeRuntimeConfig',
-      'createRuntimeIdentity',
-      'createSessionExecutorIdentity',
-      'createSubprocessCommandExecutor',
-      'createSameProcessEvaluatorAdapter',
-      'createSameProcessExecutorAdapter',
-      'rubricJudgeInstrumentId',
-      'runEvaluation',
-      'runExecutorConformance',
-    ],
-    types: [
-      'CreateEvaluationRuntimeInput',
-      'CreateExactMatchEvaluatorInput',
-      'CreateExecutorFnAdapterInput',
-      'CreateJsonExecutorAdapterInput',
-      'CreateJsonSessionExecutorAdapterInput',
-      'CreateRubricJudgeKitInput',
-      'CreateRubricJudgeEvaluatorInput',
-      'CreateSameProcessEvaluatorAdapterInput',
-      'CreateSameProcessExecutorAdapterInput',
-      'CreateSubprocessCommandExecutorInput',
-      'EvaluationEventObserver',
-      'EvaluationRuntimeSupportPorts',
-      'EvaluationRuntimeTarget',
-      'ExecResult',
-      'ExactMatchDefinitionBuilderInput',
-      'ExactMatchTarget',
-      'ExecutorFn',
-      'ExecutorFnInputMapper',
-      'ExecutorFnResultMapper',
-      'ExecutorInput',
-      'ExecutorConformanceProbeInput',
-      'ExecutorConformanceResult',
-      'InvokeExecutorIdentityDeclaration',
-      'JsonExecutorInvocation',
-      'JsonExecutorInvocationResult',
-      'JsonExecutorSession',
-      'JsonSessionExecutorAttempt',
-      'JsonSessionExecutorContext',
-      'MeasurementEventDeliveryInput',
-      'MeasurementFailurePolicyInput',
-      'MeasurementAttemptBudgetScopeInput',
-      'MeasurementBudgetPolicyInput',
-      'MeasurementBudgetScopeInput',
-      'MeasurementCachePolicyInput',
-      'MeasurementEvidencePolicyInput',
-      'MeasurementPolicyBuilderInput',
-      'MeasurementProviderCostLimitInput',
-      'MeasurementRetryBackoffInput',
-      'MeasurementRetryPolicyInput',
-      'MeasurementRunBudgetScopeInput',
-      'MeasurementStagePolicyInput',
-      'OmkLlmJudgeEffort',
-      'OmkLlmJudgeInvocationPort',
-      'OmkLlmJudgeInvocationRequest',
-      'OmkLlmJudgeInvocationResult',
-      'PairedComparisonDefinitionBuilderInput',
-      'RubricJudgeEvaluatorBinding',
-      'RubricJudgeEvaluatorDefinitionBuilderInput',
-      'RubricJudgeKit',
-      'RunEvaluationInput',
-      'RuntimeConformanceCheck',
-      'RuntimeIdentityDeclaration',
-      'SessionExecutorIdentityDeclaration',
-      'SubprocessCommandConfiguration',
-      'SubprocessCommandValueParser',
-      'RuntimePortRegistration',
-      'RuntimeValueParser',
-      'SameProcessEvaluatorImplementation',
-      'SameProcessExecutorImplementation',
-      'SameProcessOperationScope',
-      'SameProcessResourceLeaseAccess',
-      'SameProcessRunScope',
-      'AllowedToolsInput',
-      'AllowedToolsPlan',
-      'McpConfigAccess',
-      'McpConfigDescriptor',
-      'McpConfigInput',
-      'McpConfigLease',
-      'McpConfigOpenRequest',
-      'McpConfigPlan',
-      'McpConfigProvider',
-      'MockInterceptionAccess',
-      'MockInterceptionDecision',
-      'MockInterceptionDescriptor',
-      'MockInterceptionLease',
-      'MockInterceptionOpenRequest',
-      'MockInterceptionProvider',
-      'MockInterceptionRequest',
-      'WorkspaceAccess',
-      'WorkspaceDescriptor',
-      'WorkspaceInput',
-      'WorkspaceLease',
-      'WorkspaceOpenRequest',
-      'WorkspacePlan',
-      'WorkspaceProvider',
-    ],
-  },
-  'eval-runtime/contracts': {
-    entry: 'contracts',
-    values: [
-      'RUBRIC_JUDGE_BINDINGS',
-      'RUBRIC_JUDGE_CONTEXT_SCHEMA',
-      'RUBRIC_JUDGE_CONTEXT_SCHEMA_VERSION',
-      'RUBRIC_JUDGE_EVALUATOR_IMPLEMENTATION_ID',
-      'RUBRIC_JUDGE_EVIDENCE_SCHEMA',
-      'RUBRIC_JUDGE_EVIDENCE_SCHEMA_VERSION',
-      'RUBRIC_JUDGE_INSTRUMENT_SCHEMA',
-      'RUBRIC_JUDGE_INSTRUMENT_SCHEMA_VERSION',
-      'SOURCE_NEUTRAL_TRACE_SCHEMA_DESCRIPTOR',
-      'SOURCE_NEUTRAL_TRACE_SCHEMA_VERSION',
-      'SOURCE_NEUTRAL_TRACE_WITHOUT_MOCKS_SCHEMA_DESCRIPTOR',
-      'SourceNeutralMockStatsSchema',
-      'SourceNeutralTraceSchema',
-      'SourceNeutralTraceWithoutMocksSchema',
-      'attachSourceNeutralMockStats',
-      'parseSourceNeutralTrace',
-    ],
-    types: [
-      'RubricJudgeConfig',
-      'RubricJudgeCriterion',
-      'RubricJudgeInstrument',
-      'RubricJudgeRuntimeConfig',
-      'RubricJudgeTracePolicy',
-      'SourceNeutralMockStats',
-      'SourceNeutralTrace',
-    ],
-  },
-} as const;
-
-function declarationExports(entry: string): { values: string[]; types: string[] } {
-  const file = resolve(`dist/eval-runtime/${entry}.d.ts`);
-  if (!existsSync(file)) throw new Error(`缺少 ${file}；请先运行 yarn build。`);
-  const source = ts.createSourceFile(
-    file,
-    readFileSync(file, 'utf8'),
-    ts.ScriptTarget.Latest,
-    true,
-    ts.ScriptKind.TS,
-  );
-  const values: string[] = [];
-  const types: string[] = [];
-  for (const statement of source.statements) {
-    if (!ts.isExportDeclaration(statement)
-        || statement.exportClause === undefined
-        || !ts.isNamedExports(statement.exportClause)) continue;
-    for (const element of statement.exportClause.elements) {
-      (statement.isTypeOnly || element.isTypeOnly ? types : values).push(element.name.text);
-    }
-  }
-  return { values: values.sort(), types: types.sort() };
-}
 
 describe('published eval-runtime API allowlist', () => {
   it('exposes the independent-group design through the public TypeScript contract', () => {
